% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediawiki_document.R
\name{mediawiki_document}
\alias{mediawiki_document}
\title{Convert to MediaWiki Markup}
\usage{
mediawiki_document(keep_md = FALSE)
}
\arguments{
\item{keep_md}{Keep the markdown file generated by knitting.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
Format for converting from R Markdown to MediaWiki Markup.
}
\seealso{
\href{https://www.mediawiki.org/wiki/Help:Formatting}{MediaWiki markup
documentation}
}
