% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdwiki-package.R
\docType{package}
\name{rmdwiki-package}
\alias{rmdwiki}
\alias{rmdwiki-package}
\title{rmdwiki: Wiki Markup Output Formats for R Markdown}
\description{
Create wiki markup output from R Markdown (rmd) documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dpprdan/rmdwiki}
  \item Report bugs at \url{https://github.com/dpprdan/rmdwiki/issues}
}

}
\author{
\strong{Maintainer}: Daniel Possenriede \email{possenriede+r@gmail.com} (\href{https://orcid.org/0000-0002-6738-9845}{ORCID})

Other contributors:
\itemize{
  \item Julian Reinhold \email{holidayjulian@mailbox.org} [contributor]
  \item Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID}) [contributor]
}

}
\keyword{internal}
