% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textile_document.R
\name{textile_document}
\alias{textile_document}
\title{Convert to Textile Markup}
\usage{
textile_document(keep_md = FALSE)
}
\arguments{
\item{keep_md}{Keep the markdown file generated by knitting.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
Format for converting from R Markdown to Textile.
}
\seealso{
\href{https://www.promptworks.com/textile}{Textile reference manual}
}
