% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwiki_document.R
\name{xwiki_document}
\alias{xwiki_document}
\title{Convert to XWiki Markup}
\usage{
xwiki_document(keep_md = FALSE, pandoc_args = NULL)
}
\arguments{
\item{keep_md}{Keep the markdown file generated by knitting.}

\item{pandoc_args}{Additional command line options to pass to pandoc.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Format for converting from R Markdown to XWiki syntax.
}
\seealso{
\href{https://www.xwiki.org/xwiki/bin/view/Documentation/UserGuide/Features/XWikiSyntax/}{XWiki reference manual}

Other wiki formats: 
\code{\link{dokuwiki_document}()},
\code{\link{mediawiki_document}()},
\code{\link{textile_document}()}
}
\concept{wiki formats}
