% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediawiki_document.R
\name{mediawiki_document}
\alias{mediawiki_document}
\title{Convert to MediaWiki Markup}
\usage{
mediawiki_document(keep_md = FALSE, pandoc_args = NULL, wikitable = TRUE)
}
\arguments{
\item{keep_md}{Keep the markdown file generated by knitting.}

\item{pandoc_args}{Additional command line options to pass to pandoc.}

\item{wikitable}{Use \href{https://www.mediawiki.org/wiki/Help:Tables#class="wikitable"}{\code{class=wikitable}}
for tables, default is \code{TRUE}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Format for converting from R Markdown to MediaWiki Markup.
}
\seealso{
\href{https://www.mediawiki.org/wiki/Help:Formatting}{MediaWiki markup documentation}

Other wiki formats: 
\code{\link{dokuwiki_document}()},
\code{\link{textile_document}()},
\code{\link{xwiki_document}()}
}
\concept{wiki formats}
